:recycle: [#756] refactor: Move strategies to separate module for reuse
